refactor(positioning): replace integer-shift with fractional NUMERIC coordinates
Implements ADR-004. Card move and column reorder now store order as a
NUMERIC coordinate (float64) and write a single row per move; inline
repository-layer rebalancing rewrites a partition to STEP multiples when a
neighbor gap drops below the threshold. Removes the integer-shift methods
(IncrementPositionsFrom, DecrementPositionsAfter, ReorderPositions,
DeleteWithReorder) and the max+1 clamp.

- position: INTEGER -> NUMERIC (migration 000006); head-inserts may be 0/negative
- STEP/RebalanceThreshold defined once in domain, shared by use case + repo
- rebalanceIfNeeded returns the moved item's post-rebalance position so the
  response never reports a value the DB no longer holds

+++ collabotask-backend/internal/delivery/http/request/card_request.go
@@ -21,10 +21,8 @@ type UpdateCardRequest struct {
 }
 
 type MoveCardRequest struct {
-	// Do Not Delete, maybe will use later
-	// FromColumnID uuid.UUID `json:"from_column_id" binding:"required"`
 	ToColumnID uuid.UUID `json:"to_column_id" binding:"required"`
-	// Pointer so presence is enforced (omitted → 400) without rejecting the
-	// legitimate value 0 (top slot). The use case validates the range.
-	ToPosition *int `json:"to_position" binding:"required"`
+	// Pointer enforces presence (omitted or null → 400). Any finite float64 value
+	// is valid; the JSON decoder rejects NaN/±Inf before binding runs.
+	ToPosition *float64 `json:"to_position" binding:"required"`
 }
