+++ collabotask-backend/internal/domain/entity/column.go
@@ -10,7 +10,7 @@ type Column struct {
 	ID        uuid.UUID `json:"id" db:"id"`
 	BoardID   uuid.UUID `json:"board_id" db:"board_id"`
 	Title     string    `json:"title" db:"title"`
-	Position  int       `json:"position" db:"position"`
+	Position  float64   `json:"position" db:"position"`
 	CreatedAt time.Time `json:"created_at" db:"created_at"`
 	UpdatedAt time.Time `json:"updated_at" db:"updated_at"`
 }
